Conflicts:
	src/STPManager/STP.cpp
	src/extlib-abc/aig.h
	src/extlib-abc/vecStr.h
	src/interface/C/c_interface.cpp
	src/sat/SATSolver.h
	src/sat/cryptominisat2/Subsumer.cpp
	src/sat/minisat/core_prop/Solver_prop.cc
	src/to-sat/BitBlaster.cpp

Manually merged a number of warning fixes that both I and
others had done in slightly different ways.

+++ src/simplifier/PropagateEqualities.cpp
@@ -31,10 +31,10 @@ namespace BEEV
 
     bool result = false;
     if (k == XOR)
-      for (int i = 0; i < lhs.Degree(); i++)
+      for (size_t i = 0; i < lhs.Degree(); i++)
         {
           ASTVec others;
-          for (int j = 0; j < lhs.Degree(); j++)
+          for (size_t j = 0; j < lhs.Degree(); j++)
             if (j != i)
               others.push_back(lhs[j]);
 
@@ -61,7 +61,7 @@ namespace BEEV
   bool
   PropagateEqualities::searchTerm(const ASTNode& lhs, const ASTNode& rhs)
   {
-    const int width = lhs.GetValueWidth();
+    const unsigned width = lhs.GetValueWidth();
 
     if (lhs == rhs)
       return true;
@@ -77,10 +77,10 @@ namespace BEEV
       return searchTerm(lhs[0], nf->CreateTerm(BVNEG, width, rhs));
 
     if (lhs.GetKind() == BVXOR || lhs.GetKind() == BVPLUS)
-      for (int i = 0; i < lhs.Degree(); i++)
+      for (size_t i = 0; i < lhs.Degree(); i++)
         {
           ASTVec others;
-          for (int j = 0; j < lhs.Degree(); j++)
+          for (size_t j = 0; j < lhs.Degree(); j++)
             if (j != i)
               others.push_back(lhs[j]);
 
@@ -306,4 +306,3 @@ namespace BEEV
     } //end of CreateSubstitutionMap()
 
 }
-;
